Mon, Nov 28, 2022, 6:06 PM +00:00

+++ recover/recover.c
@@ -26,7 +26,7 @@ int main(int argc, char *argv[])
 
     //create buffer
     int blockSize = 0x200;
-    int buffer[512];
+    BYTE buffer[512];
 
     //store file name and jpeg count
     char *filename = malloc(8 * sizeof(char));
